計器(sim): 解けなかった連立の反復の番号の型を、反復回数と同じ型へ揃える
位置の反復回数が返す型と細分の記録が持つ番号の型が違い、型の食い違いで落ちた。数値の
変換で合わせず、記録の側を反復回数と同じ型にした。

+++ crates/blitz_sim/src/contact/scene/numeric_contract_record/tolerance_dependence/substep_record.rs
@@ -23,7 +23,7 @@ pub(super) struct 許容差依存の細分の記録 {
     pub(super) 入口: Option<第一反復の入口の読み取り>,
     pub(super) 解いた回数: 部分集合を解いた回数,
     pub(super) 候補一と勝った有効集合が違うか: Option<bool>,
-    pub(super) 解けなかった連立: Vec<(usize, 解けなかった連立の読み取り)>,
+    pub(super) 解けなかった連立: Vec<(u32, 解けなかった連立の読み取り)>,
 }
 
 impl 許容差依存の細分の記録 {
